dispose context if not added from dependency injection

+++ test/nunittest/seed/AppDbContextBase.cs
@@ -6,6 +6,7 @@ namespace nunittest.seed;
 
 public class AppDbContextBase : IDisposable
 {
+    private bool disposedvalue;
     protected readonly InventoryDbContext _context;
 
     public AppDbContextBase()
@@ -25,9 +26,22 @@ public class AppDbContextBase : IDisposable
         CdcDbInitializer.Initialize(_context);
     }
 
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
+        {
+            if (_context != null)
+            {
+                _context.Database.EnsureDeleted();
+                _context.Dispose();
+            }
+
+            disposedvalue = true;
+        }
+    }
     public void Dispose()
     {
-        _context.Database.EnsureDeleted();
-        _context.Dispose();
+        Dispose(disposing: true);
+        GC.SuppressFinalize(this);
     }
 }
